server/app/requirements.txt
bcrypt==4.1.3
fastapi==0.111.0
fastapi-jwt-auth==0.5.0
passlib==1.7.4
pydantic==2.7.4
python-dotenv==1.0.1
supabase==2.5.1
uvicorn==0.30.1
